Reson Filter (Just for recorded performance)
Implemented a reson filter that is applied only to a recorded performance, and that can be dynamically updated. (Note: This does not include XML integration yet. Also, the fuzz effect was implemented in a previous version, but I don't think I documented that at the time since I hadn't tested it yet, but it works, too.)