Applying Dynamic Effects to Recording - XML Parsing
The XML parsing system now allows for the dynamic changing of effects for the recorded performance. Specifically, as children of the recording node, add an "effectrecorded" child, and specify the measure and beat when it will apply, along with an integer to represent the effect and a double to represent the parameter. For the integers representing the effect function calls, 0 sets the amplitude, 1 sets the fuzz, 2 enables or disables the dynamic reson filter (0.0 to disable, any other value to enable), 3 sets the frequency of the dynamic reson filter, and 4 sets the band of the dynamic reson filter.